One time Map Generation
hoping this structure makes sense. I copied the TicTacToe example and made everything a class extending React.Component.

the App opens a GameBoard which does all the work.

Has a major quirk, I can't access the saved board for some reason. It is null. other things seem to work (click statement for example), but that does not.